progress on project rubric refactor

+++ src/Controller/ProjectController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Project;
 use App\Form\ProjectType;
 use App\Repository\ProjectRepository;
+use App\Service\Permissions;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,19 +19,36 @@ class ProjectController extends AbstractController
 {
 
     /**
-     * @Route("/{labelsetid}/new", name="project_new", methods={"GET","POST"})
+     * @Route("/{courseid}/index", name="project_index", methods={"GET"})
      */
-    public function new(Request $request, $labelsetid): Response
+    public function index(ProjectRepository $projectRepository, Permissions $permissions, string $courseid): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_INSTRUCTOR');
+        $role = $permissions->getCourseRole($courseid);
+        $course = $this->getDoctrine()->getManager()->getRepository('App:Course')->findOneByCourseid($courseid);
+        return $this->render('project/index.html.twig', [
+            'projects' => $projectRepository->findByCourse($courseid),
+            'course' => $course,
+            'role' => $role
+        ]);
+    }
+    /**
+     * @Route("/{courseid}/new", name="project_new", methods={"GET","POST"})
+     */
+    public function new(Request $request, $courseid): Response
+    {
+        $this->denyAccessUnlessGranted('ROLE_INSTRUCTOR');
+
         $username = $this->getUser()->getUsername();
         $user = $this->getDoctrine()->getManager()->getRepository('App:User')->findOneByUsername($username);
-        $options = ['user' => $user];
+        $course = $this->getDoctrine()->getManager()->getRepository('App:Course')->findOneByCourseid($courseid);
+        $options = ['user' => $user, 'courseid' => $courseid];
         $project = new Project();
         $form = $this->createForm(ProjectType::class, $project, ['options' => $options]);
         $form->handleRequest($request);
         $entityManager = $this->getDoctrine()->getManager();
-        $labelset = $this->getDoctrine()->getManager()->getRepository('App:Labelset')->findOneById($labelsetid);
-        $project->setLabelset( $labelset);
+
+        $project->setCourse($course);
         $project->setUser($user);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -38,7 +56,7 @@ class ProjectController extends AbstractController
             $entityManager->persist($project);
             $entityManager->flush();
             $this->addFlash('notice', 'Your Project has been created.');
-            return $this->redirectToRoute('labelset_show', ['id'=> $labelsetid]);
+            return $this->redirectToRoute('course_show', ['courseid'=> $courseid]);
         }
 
         return $this->render('project/new.html.twig', [
@@ -52,20 +70,28 @@ class ProjectController extends AbstractController
      */
     public function edit(Request $request, Project $project): Response
     {
-        if (!($this->getUser() == $project->getUser() or $this->isGranted('ROLE_ADMIN'))) {
-            throw new AccessDeniedException('You do not have permissions to do this!');
-        }
+
+        $this->denyAccessUnlessGranted('ROLE_INSTRUCTOR');
+
         $username = $this->getUser()->getUsername();
         $user = $this->getDoctrine()->getManager()->getRepository('App:User')->findOneByUsername($username);
-        $labelset = $project->getLabelset();
-        $options = ['user' => $user];
+
+        if ($project->getLabelset()) {
+            $courses = $project->getLabelset()->getCourses();
+            $courseid = $courses[0]->getId();
+        }
+        else {
+            $courseid = $project->getCourse()->getId();
+        }
+
+        $options = ['user' => $user, 'courseid' => $courseid];
         $form = $this->createForm(ProjectType::class, $project, ['options' => $options]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
             $this->addFlash('notice', 'Your Project has been updated.');
-            return $this->redirectToRoute('labelset_show', ['id'=> $labelset->getId()]);
+            return $this->redirectToRoute('course_show', ['courseid'=> $courseid]);
         }
 
         return $this->render('project/edit.html.twig', [
@@ -79,15 +105,16 @@ class ProjectController extends AbstractController
      */
     public function delete(Request $request, Project $project): Response
     {
-        $labelset = $project->getLabelset();
+        $this->denyAccessUnlessGranted('ROLE_INSTRUCTOR');
 
+        $courseid = $project->getCourse()->getId();
         if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($project);
             $entityManager->flush();
         }
         $this->addFlash('notice', 'Your Project has been deleted.');
-        return $this->redirectToRoute('labelset_show', ['id'=> $labelset->getId()]);
+        return $this->redirectToRoute('project_index', ['courseid'=> $courseid]);
     }
 }
